Rename `Section` to `Grid` and `Container` to `Section`.

+++ Sources/Ignite/Styles/SpacingAmount.swift
@@ -7,7 +7,7 @@
 
 /// Adaptive spacing amounts that are used by Bootstrap to provide consistency
 /// in site design.
-public enum SpacingAmount: Double {
+public enum SpacingAmount: Int {
     case none = 0
     case extraSmall
     case small
